+++ src/ploneintranet/docconv/client/__init__.py
@@ -1,9 +1,22 @@
-from Products.CMFCore.utils import getToolByName
+# from Products.CMFCore.utils import getToolByName
 from zope.component.hooks import getSite
+from collective.documentviewer.settings import GlobalSettings
 
 from logging import getLogger
 
 logger = getLogger(__name__)
+
+
+# documentviewer does set the layout in two places. One can be turned off
+# the following one can't. So we have to patch.
+def _handle_layout(self):
+    """
+    Deactivate the layout setting part of documentviewer as we want our own
+    """
+    pass
+
+from collective.documentviewer.convert import Converter
+Converter.handle_layout = _handle_layout
 
 
 def initialize(context):
@@ -11,7 +24,9 @@ def initialize(context):
 
 
 def is_autoconv_enabled():
-    site = getSite()
-    portal_properties = getToolByName(site, 'portal_properties')
-    site_properties = portal_properties.site_properties
-    return site_properties.getProperty('docconv_autoconv', False)
+    """
+    if enabled, viewing items without a preview will trigger generation
+    of a preview
+    """
+    gsettings = GlobalSettings(getSite())
+    return gsettings.auto_convert